Support json_feed, json_simple, and rss news formats (#49)
You can now specify the output format for the news scraper with the `--format` option, which can be:
- `json_simple` (default) the current format
- `json_feed` the JSONFeed standard (https://jsonfeed.org/version/1)
- `rss` for RSS 2.0

Eventually, we'll switch `json_feed` to be the default, but for now we need to not break existing automation.

This uses dataclasses for some of its work, which are new in Python 3.7. I think that’s our minimum version, so this should be OK. (Is that right, @ldtcooper?)

You can also specify more than one format, and specify an output directory. That way, a command like this:

```sh
$ python scraper_news.py san_francisco alameda --format json_feed --format rss --output feed_data
```

Can write out 4 files, which would be a pain to deal with on STDOUT:
- `feed_data/san_francisco.json`
- `feed_data/san_francisco.rss`
- `feed_data/alameda.json`
- `feed_data/alameda.rss`

Fixes #40, fixes #42.