Ensure order of items in news feeds is stable
In situations where multiple feed items have the same publish time (usually because the county only specifies the date and not the time), we sometimes wind up with feed items getting a slightly different order between scraper runs. We now use the ID of the news items as an additional sort criteria to make sure they stay in the same order.

This change also adds some actual tests!

Fixes #88.