Default to *all* counties for news scraper CLI (#94)
If you didn't specify any county names to scrape, the news scraper used to only output news for San Francisco. It now outputs news for all counties instead, which is a more sensible default and matches up with the data side. (Code that depended on the old default has now all been fixed, so this change is finally safe.)

Fixes #92.