Script to parse FAQ Content sheet into JSON
Rob's review notes taken in. Only skipped moving the questions list ('questions_list = [] # Create a list to contain questions') outside the loop - hope I understood that right! - because it causes every Question object to show up in every Section object. Also still looking at DictReader . . . not yet certain this will have much advantage given the format of the spreadsheet and of the parsed json we want (a UX/front end discrepancy that might be a lot of bother to resolve outside the code . . . still looking at that, though, and super-open to ideas :-)