Load more news from Solano via fancy XHR endpoint (#120)
The main news page for Solano County presents only a limited number of items, and doesn't give us enough to actually cover the full time range we're looking for in a lot of cases. Instead of loading and parsing the main news page, we now load the same endpoint that is used to display the county's full news archives.

NOTE: this required some changes to response body decoding since the endpoint does not specify the encoding it uses. I had to add a way to set a hard-coded fallback encoding when one cannot be inferred from the HTTP response.