Add Dockerfile that works for development use (#111)
Since it can be tough to get all the pieces installed and working right when using WSL on Windows, this dockerfile provides a way to develop and run the app in a container with all the necessary dependencies.

To use it, run any of our existing scripts through `run_docker.sh`. For example, to run the data scraper:

    $ ./run_docker.sh python scraper_data.py

This is a shortcut for two steps. First, build an image from it with:

    $ docker build -t data-covid19-sfbayarea .

And then run it via:

    # Get a bash prompt inside the container:
    $ docker run -it data-covid19-sfbayarea

    # Or instead just run the scraper in the
    # container and quit the container when the
    # scraper finishes:
    $ docker run -it data-covid19-sfbayarea python scraper_data.py --more args --here