Handle stripping whitespace better in Solano news
We previously used `element.get_text(strip=True)` when extracting text for Solano news, but that causes real issues when there are child elements (links, bold, etc.) mixed in, because it strips each bit of text independently, leaving us with no spaces at all between words that are at the edge of elements. This handles the whitespace a little better, and more like it normally gets handled in HTML.

Fixes #96.