Fix repeated news entries for Contra Costa county (#125)
Contra Costa changed the format of its news page slightly, making the table of contents entries into headers and switching to a structure where every two months shared a single parent `<div>`. For example:

    <div>
        <h3>September 2020</h3>
        <ul>[news items in here]</ul>
        <h3>August 2020</h3>
        <ul>[news items in here]</ul>
    </div>

Previously, there was one `<div>` that wrapped each heading + news list pair. The new arrangement caused us to look at all the `<li>`s for both months when we hit September and again when we hit August, because we’d look for all the `<li>`s in the month heading's parent element. Instead, we now just look for the next list after the month heading.

Fixes #123.