HOTFIX: Alameda news page now has more `<br>` (#126)
Previously, we only saw `<br>` elements on the Alameda news page as delimiters between the title and summary of a news item, or double `<br>`s between news items. Now they're all over the place, including between the date and title. Since we previously assumed the `<br>` must be separating the title and summary, that meant we kicked into summary parsing mode without ever reading any title data. Luckily, that would cause the scraper to fail rather than output bad data (all news items MUST have titles), but that still means the scraper was broken.

The new approach here is to check to see whether we've gathered any title data yet when we hit a `<br>`. If not, keep looking for the title. Otherwise, treat it like before and start reading subsequent text as the summary.