Add tests to validate output format
It turns out several of our scrapers are not quote conforming to the data model. We can and should have a class that models the data and can enforce some formatting and validation, but it's also good to test that the final result adheres to a schema, which is what this does.

Because it's tough to mock responses in Selenium, these tests run against live county websites right now. That can make them pretty slow or intense to run, so this class of tests is disabled by default and you have to opt-in to run them (in CI, we opt-in). To do so, set the `LIVE_TESTS` environment variable to `*` to run all live tests, or set it to a comma-separated list of county names to run just those counties:

    # Run live tests against all county websites.
    $ LIVE_TESTS='*' python -m pytest -v .

    # Run live tests against only San Francisco and Sonoma counties.
    $ LIVE_TESTS='san_francisco,sonoma' python -m pytest -v .