Add tests to validate output format (#130)
It turns out several of our scrapers don't quite conform to the data model. We should probably have some classes that model the data and can enforce some formatting and validation, but it's also good to test that the final result adheres to a schema, which is what this does.

Because it's tough to mock responses in Selenium, these tests run against live county websites right now. That can make them pretty slow or intense to run, so this class of tests is disabled by default and you have to opt-in to run them (in CI, we opt-in). To do so, set the `LIVE_TESTS` environment variable to `*` to run all live tests, or set it to a comma-separated list of county names to run just those counties:

    # Run live tests against all county websites.
    $ LIVE_TESTS='*' python -m pytest -v .

    # Run live tests against only San Francisco and Sonoma counties.
    $ LIVE_TESTS='san_francisco,sonoma' python -m pytest -v .

Finally, **these tests skip any county that raises an error while scraping.** Because changes on county websites frequently cause scrapers to fail, we don’t want that to create flaky, failure-prone tests here. (For example, if Alameda and Solano are both failing, we don’t want to be bothered with Alameda errors on the tests for the PR that fixes Solano.)

The validations use [Voluptuous](https://github.com/alecthomas/voluptuous), which is a reasonably nice and concise language for writing schemas. A more verbose, but more standard alternative might be to use JSONSchema.

Tests will now output useful information about what’s wrong with the output data format.