Add hospitalization data scraper script (take two) (#131)
This makes the tools in `hospitals.py` available as a CLI script. There have been some changes in the `hospitals.py` module itself, but most of the changes are just reorganization for flow and readability. The two important changes are:

* The `series` object in the JSON output is now a dict keyed by county name, to keep consistency with the other data structures;

* The `get_timeseries()` function will pull down _all_ the hospitalization data from the API (for all California counties), and then restructure and filter the output locally rather than filtering by county through the API call.

This latter change was made primarily because I'm a little uneasy about the API's filtering just using the `q` param, and haven't found a way to leverage the CKAN API SQL syntax (the URL encoding makes `SELECT * ...` it a little tricky, although I'm probably just overlooking some escape characters). 

In short, filtering locally seems simpler and more reliable, but as the data grow, we may want/need to find a way to filter more effectively both by county name and by date range.