Update date prefix removal in San Mateo news
Many news items have their title prefixed with a date in San Mateo, which can be a bit confusing and redundant, especially when we normally display the date right before the title. To allevaiate this, we attempt to remove the date at the start of the title. However, the county recently changed its date format to include abbreviations, which we didn't previously account for. It also turns out that dates are getting formatted differently in Spanish (Chinese articles seem to have no date prefix, and I don't have examples of any other languages to test).

Additionally, it turns out news items with summaries have a similar date prefix on the summary, too! We now remove it as well.

Fixes #134.