HOTFIX: Handle messy links in Alameda news titles (#141)
The Alameda news page now has links that take up only part of the title. The parser sometimes identifies these as links to language-specific versions of the news item, which causes it to stop parsing the title and start looking for language links. However, these links aren't language links, and we need to continue parsing the title as we move through them.

This fixes the issue by being *much* more specific about what we expect for the text of a language link. This may have issues if there are ever typos in the language names, but seems to work well enough.

Fixes #140.