Sonoma news: handle whitespace in tags correctly
Some of the HTML elements we transform into tags for news items in Sonoma County have line breaks in them, and they weren't being handled well. We already have tools for doing this better, but we just weren't using them. This switches over to them.

For example, in this PR: https://github.com/sfbrigade/stop-covid19-sfbayarea/pull/497

We have a news item like:

    {
      "title": "COVID-19 Rental Assistance\u00a0For Sonoma County Renters and Landlords",
      "date_published": "2020-10-14T00:00:00-07:00",
      ...
      "tags": [
        "Sonoma CountyCommunity Development Commission"
      ]
    }

Note that there is not space between "Sonoma County" and "Community" in the tag name. The HTML for this looks like:

    <div class="source">
      Sonoma County <br />
      Community Development Commission
    </div>

So we just need to handle the whitespace between text nodes a little better here.