HOTFIX: Handle Alameda news with more nested nodes
Alameda news items used to follow a pretty standard pattern where none of the components are contained in other elements. It typically looked like:

    <strong>Date</strong> <a href="...">Title</a>

But now some parts of some news items are getting wrapped in `div` elements, causing us to have trouble iterating through all the parts and extracting the complete data. For example, a news item on the page now looks like this:

    <div><strong>Date</strong><div> <a href="...">Title</a>

Note this isn't really consistent across news items, so it's all pretty messy. In any event, the new structure causes a couple problems that get worked around in this change.