Rewrite marin.get_chart_meta to load page once
The metadata method for Marin county reloaded the page repeatedly and unnecessarily, making things tough on us and tough on Marin's web servers, too. This updates it to only load the page once.

While I was at it, I also updated it to preserve the order of the metadata and added some other notes about code I didn't understand here.