Remove print statements causing output errors
We should not be printing anything from non-CLI related code (or, at least, not to stdout, which is where data, not errors go). This error message was printing in front of the output JSON data, and causing it not to parse. Since it failed that early, our actual error reporting also broke (because it expected to be able to parse the output as JSON).