HOTFIX: Handle strings in Alameda %-positive tests
For some days in Alameda, we get strings instead of numbers in the percent positive tests timeseries. This fix is not particularly generic, but solves the immediate problem. It looks like PowerBI may try to preserve precision by sending strings when float64 cannot accurately represent a value. We'll have to think about more generic ways to account for this.