HOTFIX: handle repeat rows in Solano age groups
The last day of Solano's data seems to be getting repeated in their dataset (every row is repeated once, totally verbatim, as far as I can tell). I'm not sure why that's happening (it's there that way in the source data if you go look on the web), but it seems benign, so I've just added a fix here to de-duplicate the age groups (we just take the last value we have for each group). This already happens naturally for race/ethnicity and gender (since we build dicts instead of lists), so there's no need to fix those.