HOTFIX: Napa all news page is broken
Napa county's page for all news is broken (`https://www.countyofnapa.org/CivicAlerts.aspx?sort=date`) and doesn't load, so we can't actually scrape it. However, if we look at news from only a subset of categories, it works fine (it appears that category 17 [Board of Supervisors] is broken). This lists all the categories except one that breaks, so the page loads and gives us most news.