HOTFIX: San Mateo tests table was renamed
It looks like someone renamed the table and columns for test data in San Mateo, which broke the scraper. I've updated with the new names and verified that the data looks at least *roughly* like it used to. (It seems the timeseries now starts a month later than it used to, though. I tried using a `Where` clause to extend it, but no dice. This is just all that's there now.)