HOTFIX: use new CHHS dataset for San Mateo deaths (#189)
San Mateo County does not provide a timeseries of deaths, so we previously used the state's data portal for that data. The state dataset we were pulling from has been deprecated in favor of one provided by CHHS, and this updates the scraper to use that. Unfortunately, the totals it provides no longer match up with the county's provided totals, but I'm not sure what to do about that at this point other than list it in the `meta` section. :(

The old dataset was at: https://data.ca.gov/dataset/covid-19-cases

Fixes #186.