HOTFIX: Get San Mateo deaths by day from LA Times
The state has ceased publishing COVID-19 deaths as a timeseries, so we need to load the data from another source. LA Times's data appears to have been more consistent with the county than the data we were previously loading from CHHS, so this not only is a hotfix, but probably improves our data quality anyway.

Fixes #190.